feat: enable clear signing (#37)
* Add ledger feature flag to isolate targets and make our crate a lib with a binary

Move main code to its own file as a separate program that depends on our library

update deps and ran crate

Delete unused files

Update main

Use ledger flag

* Add review transaction instruction and handlers

* Implement hashing for transaction

* Add helper method to get full flash buffer to parser transaction

* Add zemu test to review transaction

* Add and fix tests for transaction merkle note and affine point parsing

* Send output to stdout in tests

* Store in RAM computed transaction hash and use it to check that received hash is correct

* add new dep

* Add ui machinery to test using rust and cargo insta(taken from avax)

* Add UI unit testing in rust for transactions

* Improve transaction review in zemu tests

* Clean ups

* Try to not use display and format which goes through a writer! macro call

* minor changes to reduce pic issues

* Update tests

* Disable hash check for now

* Remove logs

* Update test and snapshots

* feat: bring ui review transaction fn from original branch

* Avoid inlining

* Fix note parser

* Use lexical core for big integer to string conversion without allocation and update shapshots

* Update transaction unit tests

* fix: remove wrong rebased code

* Fix transaction version values

* Fix parser and transaction hash for mints and spends

* Enable hash check

* feat: enable reviewTx on stax and flex devices (#39)

* feat: refactor ui_review

- remove format
- rework review messages params

* feat: ger rid of format fn

* feat: enable reviewTx on stax and flex

* fix: issue

* feat: increase dkg keys to move it to a different section on binary

* feat: add snapshots

* feat: adjust review flow to fit stax and flex

* feat: update nanosp snapshots

* feat: update snapshots

* feat: update snapshots

* feat: update snapshots

* feat: update snapshots

* fix: format

* Check transaction was approved and compare hashes before starting the signing process

* Add transaction review before signing

* update snapshots

* Removed logs

* Fix warnings wip

* feat: enable stax and flex devices

* feat: update snapshots

* fix: stax and flex

* feat: update snapshots

---------

+++ app/src/handlers/dkg_backup_keys.rs
@@ -35,7 +35,7 @@ pub fn handler_dkg_backup_keys(comm: &mut Comm, ctx: &mut TxContext) -> Result<(
     let min_signers = DkgKeys.load_min_signers()?;
     let participants = DkgKeys.load_identities()?.len();
 
-    if !ui_review_backup_keys(public_address, participants, min_signers)? {
+    if !ui_review_backup_keys(public_address, participants as u8, min_signers as u8)? {
         return Err(AppSW::Deny);
     }
